Adding windows install file
This file contains a short description on how to compile
the script on a windows computer. Windows does not
support makefiles. Therefore the user is required to do
the compilation on his own.

This file also covers on how to install sphinx because it
is not part of every python package.